Create Unwanted Ape Club Submitting a branding request
What is the Unwanted Ape Club and what are Unwanted Ape Club NFTs?

The Unwanted Ape Club (UAC) is an ambitious community project of 4,444 unique NFTs. Each UAC NFT is an original piece of digital art, created without using the feature combinations of the famous Ethereum BAYC. Created by an enthusiast, UAC was born out of a desire to prove that anyone can create a complete NFT collection with few resources, patience, motivation, and extensive research.